requirements.txt
appdirs==1.4.3
APScheduler==3.6.3
apturl==0.5.2
asgiref==3.4.0
asyncio==3.4.3
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.4
dominate==2.6.0
duplicity==0.8.12.0
entrypoints==0.3
fastapi==0.66.0
fasteners==0.14.1
filelock==3.0.12
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.15.1
future==0.18.2
h11==0.12.0
httplib2==0.14.0
httptools==0.2.0
idna==2.10
importlib-metadata==1.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pipenv==11.9.0
protobuf==3.6.1
pyaes==1.6.1
pyasn1==0.4.8
pycairo==1.16.2
pycups==1.9.73
pydantic==1.8.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
pyTelegramBotAPI==3.7.6
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.4
python-dotenv==0.15.0
python-telegram-bot==13.7
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
redis==3.5.3
reportlab==3.5.34
requests==2.25.1
requests-unixsocket==0.2.0
rsa==4.7
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.2.4
starlette==0.14.2
systemd-python==234
Telethon==1.19.0
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.2
usb-creator==0.3.7
uvicorn==0.14.0
uvloop==0.15.2
virtualenv==20.0.17
virtualenv-clone==0.3.0
visitor==0.1.3
wadllib==1.3.3
watchgod==0.7
websockets==9.1
Werkzeug==2.0.1
WTForms==2.3.3
xkit==0.0.0
zipp==1.0.0
